dev_scripts/requirements.txt
ctransformers
faiss-cpu
langchain==0.0.331
pypdf	
python-box
sentence-transformers
chainlit
pydantic==1.10.8